website でも共有した textlint を使うようにして
