--- packages/website/.babelrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  plugins: ['istanbul'],
-  presets: [
-    [
-      '@babel/preset-env',
-      {
-        corejs: 3,
-        useBuiltIns: 'entry'
-      }
-    ],
-    '@babel/preset-react',
-    '@babel/preset-typescript'
-  ]
-};
