+++ packages/website/package.json
@@ -61,7 +61,6 @@
     "kuromoji": "^0.1.2",
     "nyc": "^15.0.0",
     "path": "^0.12.7",
-    "raw-loader": "^4.0.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-helmet": "^6.0.0",
