--- packages/website/src/dict/index.ts
@@ -1,3 +1,0 @@
-import initializeDict from './initializeDict';
-
-export default initializeDict;
