+++ packages/website/src/index.tsx
@@ -13,6 +13,7 @@ import {
   Theme,
   ThemeProvider as MuiThemeProvider,
 } from '@material-ui/core/styles';
+import * as BrowserFS from 'browserfs';
 import * as Sentry from '@sentry/browser';
 import App from './App';
 
@@ -32,6 +33,30 @@ if (process.env.NODE_ENV === 'production') {
     window.addEventListener('load', () => navigator.serviceWorker.register('service-worker.js'));
   }
 }
+
+declare global {
+  interface Window {
+    kuromojin?: {
+      dicPath?: string;
+    };
+  }
+}
+
+window.kuromojin = {
+  dicPath: 'dict',
+};
+
+BrowserFS.install(window);
+BrowserFS.configure(
+  {
+    fs: 'LocalStorage',
+  },
+  (exception) => {
+    if (exception) {
+      throw exception;
+    }
+  }
+);
 
 const theme = createMuiTheme({
   palette: {
